move workflowExecutor to new file
