+++ javascript/workflow/src/resources/utils.ts
@@ -3,6 +3,8 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 import type { ResourceSchema } from '@datashaper/schema'
+import { all, op } from 'arquero'
+import type ColumnTable from 'arquero/dist/types/table/column-table.js'
 import fetch from 'cross-fetch'
 
 export const isWorkflow = (r: ResourceSchema): boolean =>
@@ -85,4 +87,15 @@ function resolveRawContent(
 	if (files?.has(resource)) {
 		return files?.get(resource)
 	}
+}
+
+export function withRowNumbers(
+	table: ColumnTable | undefined,
+): ColumnTable | undefined {
+	return table?.derive(
+		{
+			index: op.row_number(),
+		},
+		{ before: all() },
+	)
 }
